unsuccessful pi integration
not seen here, as it has been edited until it was back to this state, was a ros serrial node that functioned when on pc, did not on pi, then did not when back on pc. a 3v3 to 5v level shifter has been implemented, able to produce a 5v pwm using the 5v source, 3v3 source, and 3v3 pwm gpio.