robot motion video
renamed video showing the motion of the robot after the project showcase. the video is not of the best quality as it was done using an ipad one-handed while the other controlled the robot via the wired remote. the remote is best controlled two-handed and stability and focus is suffered in the video capture itself.